k reorder deterministic now; found ism is not deterministic for bigblue4

+++ dreamplace/ops/global_swap/src/global_swap_cuda_kernel.cu
@@ -88,7 +88,7 @@ struct SwapState
 
     SwapCandidate<T>* candidates; 
 
-    long* net_hpwls; ///< HPWL for each net, use integer to get consistent values 
+    double* net_hpwls; ///< HPWL for each net, use integer to get consistent values 
     unsigned char* node_markers; ///< markers for cells 
 
     int batch_size;
